// core/libs/editor/src/lib/editor-toolbar/interfaces/export-package.ts
/*
 * Copyright (c) 2023 Robert Bosch Manufacturing Solutions GmbH
 *
 * See the AUTHORS file(s) distributed with this work for
 * additional information regarding authorship.
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at https://mozilla.org/MPL/2.0/.
 *
 * SPDX-License-Identifier: MPL-2.0
 */

import {ValidFile} from './valid-file';
import {MissingElement} from './missing-element';

export interface ExportPackage {
  invalidFiles: string[];
  validFiles: ValidFile[];
  missingElements: MissingElement[];
}