// core/libs/editor/src/lib/editor-toolbar/interfaces/valid-file.ts
/*
 * Copyright (c) 2022 Robert Bosch Manufacturing Solutions GmbH
 *
 * See the AUTHORS file(s) distributed with this work for
 * additional information regarding authorship.
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at https://mozilla.org/MPL/2.0/.
 *
 * SPDX-License-Identifier: MPL-2.0
 */

import {ViolationReport} from './violation-report';

export interface ValidFile {
  aspectModelFileName: string;
  violationReport: ViolationReport;
  fileAlreadyDefined: boolean;
}